+++ R/trim-osmdata.R
@@ -16,7 +16,12 @@
 #'
 #' @note It will generally be necessary to pre-load the \pkg{sf} package for
 #' this function to work correctly.
+#' @note Caution is advised when using polygons obtained from Nominatim via 
+#' `getbb(..., format_out = "polygon"|"sf_polygon")`. These shapes can be
+#' outdated and thus could cause the trimming operation to not give results
+#' expected based on the current state of the OSM data.
 #'
+#' @family transform
 #' @export
 #' @examples
 #' \dontrun{
@@ -61,6 +66,10 @@ trim_osmdata_sfp <- function (dat, bb_poly, exclude = TRUE) {
         bb_poly <- bb_poly_to_mat (bb_poly)
 
     if (nrow (bb_poly) > 1) {
+
+        srcproj <- .lonlat() #"+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+        crs <- .sph_merc() # "+proj=merc +a=6378137 +b=6378137"
+        bb_poly <- reproj::reproj (bb_poly, target = crs, source = srcproj) [, 1:2]
 
         dat <- trim_to_poly_pts (dat, bb_poly, exclude = exclude) %>%
             trim_to_poly (bb_poly = bb_poly, exclude = exclude) %>%
@@ -131,7 +140,11 @@ trim_to_poly_pts <- function (dat, bb_poly, exclude = TRUE) {
 
     if (is (dat$osm_points, "sf")) {
 
+        srcproj <- .lonlat() #"+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+        crs <- .sph_merc() # "+proj=merc +a=6378137 +b=6378137"
+
         g <- do.call (rbind, dat$osm_points$geometry)
+        g <- reproj::reproj (g, target = crs, source = srcproj)
         indx <- sp::point.in.polygon (g [, 1], g [, 2],
                                       bb_poly [, 1], bb_poly [, 2])
         if (exclude)
@@ -159,10 +172,14 @@ trim_to_poly_pts <- function (dat, bb_poly, exclude = TRUE) {
 #' @noRd
 get_trim_indx <- function (g, bb, exclude) {
 
+    srcproj <- .lonlat() #"+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
+    crs <- .sph_merc() # "+proj=merc +a=6378137 +b=6378137"
+
     indx <- lapply (g, function (i) {
 
                         if (is.list (i)) # polygons
                             i <- i [[1]]
+                        i <- reproj::reproj (as.matrix (i), target = crs, source = srcproj)
                         inp <- sp::point.in.polygon (i [, 1], i [, 2],
                                                      bb [, 1], bb [, 2])
                         if ((exclude & all (inp > 0)) |